Refactored Products.Controller.ts, updated domain/persistence ts's and added resolve/rejects
Firstly Products.controller has been moved then simplified since most of its code is repeated in Domain/Persistence ts's.  Domain ts's are now passed with a interface to keep it separate from the rest of the code. And finally added/changed ProductsRespository to contain resolves and reject errors with HttpExecption that was created previously.